+++ lib/argon2-opencl/processingunit.cpp
@@ -8,7 +8,7 @@
 namespace argon2 {
 namespace opencl {
 
-static bool isPowerOfTwo(std::uint32_t x)
+static bool isPowerOfTwo(std::size_t x)
 {
     return (x & (x - 1)) == 0;
 }
@@ -75,7 +75,7 @@ ProcessingUnit::ProcessingUnit(
 #endif
 
         float bestTime = std::numeric_limits<float>::infinity();
-        for (std::uint32_t jpb = 1; jpb <= runner.getMaxJobsPerBlock();
+        for (std::size_t jpb = 1; jpb <= runner.getMaxJobsPerBlock();
              jpb *= 2)
         {
             float time;
